# Conflicts:
#   frontend/src/AppRouter.tsx
#   frontend/src/navigation/MainNavigator.tsx
#   frontend/src/screens/diary/DiaryEditScreen.tsx
#   frontend/src/screens/library/LibraryScreen.tsx
#   frontend/src/screens/profile/GoalListScreen.tsx

+++ backend/src/main/java/com/ssafy/loveledger/domain/account/service/AccountService.java
@@ -12,12 +12,15 @@ import com.ssafy.loveledger.domain.account.presentation.dto.response.DailyStatis
 import com.ssafy.loveledger.domain.account.presentation.dto.response.HistoryDetailResponse;
 import com.ssafy.loveledger.domain.account.presentation.dto.response.HistoryResponse;
 import com.ssafy.loveledger.domain.account.presentation.dto.response.MemberInfoResponse;
+import com.ssafy.loveledger.domain.account.presentation.dto.response.MonthlyStatisticsResponse;
 import com.ssafy.loveledger.domain.account.presentation.dto.response.SSAFYResponse;
 import com.ssafy.loveledger.domain.account.presentation.dto.response.WeekStatisticsResponse;
 import com.ssafy.loveledger.domain.history.domain.History;
 import com.ssafy.loveledger.domain.history.domain.repository.HistoryRepository;
 import com.ssafy.loveledger.domain.statistics.domain.Category;
 import com.ssafy.loveledger.domain.user.domain.User;
+import com.ssafy.loveledger.global.response.exception.ErrorCode;
+import com.ssafy.loveledger.global.response.exception.LoveLedgerException;
 import com.ssafy.loveledger.global.util.OpenFeignUtil;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -50,11 +53,16 @@ public class AccountService {
     @Value("${ssafy.apikey}")
     private String apiKey;
 
-    @Transactional(readOnly = true)
+    @Transactional//(readOnly = true)
     public Page<HistoryDetailResponse> getAccountHistory(User user, int year, int month, int day,
         int size, int pageno, String sort) {
 
-        Direction direction = sort.equals("asc") ? Direction.ASC : Direction.DESC;
+        if (user == null) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+        updateListOfHistory(user);
+
+        Direction direction = sort.equalsIgnoreCase("ASC") ? Direction.ASC : Direction.DESC;
         Pageable pageable = PageRequest.of(pageno - 1, size, Sort.by(direction, "createdTime"));
 
         Page<History> historyPage = historyRepository.findByAccountAndCreatedDate(
@@ -67,7 +75,7 @@ public class AccountService {
                 .time(history.getCreatedTime())
                 .remittance(history.getTransactionType() < 3)
                 .targetName(history.getTransactionTarget())
-                //.CategoryName(history.getCategoryId())
+                .CategoryName(history.getCategory().getName())
                 .afterAmount(history.getAmountAfterTransaction())
                 .amount(history.getTransactionAmount())
                 .build()
@@ -77,32 +85,57 @@ public class AccountService {
         return new PageImpl<>(response, pageable, historyPage.getTotalElements());
     }
 
-    @Transactional(readOnly = true)
+    @Transactional//(readOnly = true)
     public List<DailyStatisticsResponse> getAccountHistoryByMonth(User user, int year, int month,
         int size, int pageno, String sort) {
-        Direction direction = sort.equals("asc") ? Direction.ASC : Direction.DESC;
-        Pageable pageable = PageRequest.of(pageno - 1, size, Sort.by(direction, "createdTime"));
+
+        if (user == null) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        } else if (user.getAccount() == null) {
+            throw new LoveLedgerException(ErrorCode.ACCOUNT_NOT_FOUND);
+        }
+
+        updateListOfHistory(user);
+
+        Direction direction = sort.equalsIgnoreCase("asc") ? Direction.ASC : Direction.DESC;
+        Pageable pageable = PageRequest.of(pageno - 1, size, Sort.by(direction, "dayId.targetDay"));
 
         YearMonth yearMonth = YearMonth.of(year, month);
         LocalDate startDate = yearMonth.atDay(1);
         LocalDate endDate = yearMonth.atEndOfMonth();
-
-        List<DailyStatisticsResponse> responses = historyRepository.findByUserAndMonth(
-            user, startDate, endDate, pageable
-        );
-
-        return responses;
+        return historyRepository.findByUserAndMonth(user, startDate, endDate, pageable);
     }
 
     @Transactional(readOnly = true)
     public List<WeekStatisticsResponse> getAccountHistoryByWeek(User user, int year, int month) {
+
+        if (user == null) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        } else if (user.getAccount() == null) {
+            throw new LoveLedgerException(ErrorCode.ACCOUNT_NOT_FOUND);
+        }
+
         YearMonth yearMonth = YearMonth.of(year, month);
         LocalDate startDate = yearMonth.atDay(1);
         return historyRepository.findWeeklyStatistics(user, year, month, startDate);
     }
 
     @Transactional
+    public List<MonthlyStatisticsResponse> getAccountHistoryByMonth(User user, int year,
+        int month) {
+        if (user == null) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+
+        return historyRepository.findMonthlyStatistics(user, year, month);
+    }
+
+    @Transactional
     public void deleteHistory(User user, String transactionId) {
+        if (user == null) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+
         History history = historyRepository.findById(transactionId).orElse(null);
         if (history != null && history.getAccount().getUser() == user) {
             history.delete();
@@ -112,6 +145,11 @@ public class AccountService {
     @Transactional
     public void updateHistoryTarget(User user, String transactionId, String accountNo,
         String updatedTargetName) {
+
+        if (user == null) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+
         Account account = accountRepository.findById(accountNo).orElse(null);
         History history = historyRepository.findById(transactionId).orElse(null);
 
@@ -124,74 +162,67 @@ public class AccountService {
     @Transactional
     public void updateListOfHistory(User user) {
         if (user == null) {
-            throw new RuntimeException("User Not Found");
-        } else {
-            String code = generateCode();
-            String apiName = "inquireTransactionHistoryList";
-            SSAFYRequestHeader header = SSAFYRequestHeader.builder()
-                .apiName(apiName)
-                .apiServiceCode(apiName)
-                .transmissionDate(code.substring(0, 8))
-                .transmissionTime(code.substring(8, 14))
-                .institutionCode("00100")
-                .fintechAppNo("001")
-                .institutionTransactionUniqueNo(code)
-                .apiKey(apiKey)
-                .userKey(user.getUserKey())
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+
+        String code = generateCode();
+        String apiName = "inquireTransactionHistoryList";
+        SSAFYRequestHeader header = createRequestHeader(user, apiName, code);
+
+        Account account = user.getAccount().get(0);
+        if (account == null) {
+            throw new LoveLedgerException(ErrorCode.ACCOUNT_NOT_FOUND);
+        }
+
+        AccountHistoryDetailRequest request = AccountHistoryDetailRequest.builder()
+            .header(header)
+            .accountNo(account.getAccountId())
+            .startDate(account.getLastUpdated().format(formatter).substring(0, 8))
+            .endDate(code.substring(0, 8))
+            .transactionType("A")
+            .orderByType("ASC")
+            .build();
+
+        SSAFYResponse response = openFeignUtil.getListOfHistory(request);
+
+        List<Map<String, Object>> rawList = (List<Map<String, Object>>) response.getResultData()
+            .get("list");
+
+        List<HistoryResponse> res = rawList.stream()
+            .map(map -> objectMapper.convertValue(map, HistoryResponse.class))
+            .toList();
+
+        for (HistoryResponse historyResponse : res) {
+            String dateTimeString =
+                historyResponse.getTransactionDate() + historyResponse.getTransactionTime();
+            LocalDateTime dateTime = LocalDateTime.parse(dateTimeString, formatter);
+
+            String transactionType = historyResponse.getTransactionTypeName();
+            int type = switch (transactionType) {
+                case "입금" -> 1;
+                case "입금(수시입출금)" -> 2;
+                case "출금" -> 3;
+                case "출금(수시입출금)" -> 4;
+                default -> 0;
+            };
+
+            History history = History.builder()
+                .transactionId(historyResponse.getTransactionUniqueNo())
+                .createdDate(dateTime.toLocalDate())
+                .createdTime(dateTime.toLocalTime())
+                .transactionAccount(historyResponse.getTransactionAccountNo())
+                .transactionTarget(historyResponse.getTransactionSummary())
+                .transactionAmount(Long.valueOf(historyResponse.getTransactionBalance()))
+                .transactionType(type)
+                .transactionTypeName(historyResponse.getTransactionTypeName())
+                .category(Category.NOT_DEFINED) // TODO : 카테고리 분류 모델 적용 할 것
+                .account(account)
+                .AmountAfterTransaction(
+                    Long.valueOf(historyResponse.getTransactionAfterBalance())
+                )
+                .memo(historyResponse.getTransactionMemo())
                 .build();
-
-            Account account = user.getAccount().get(0);
-            AccountHistoryDetailRequest request = AccountHistoryDetailRequest.builder()
-                .header(header)
-                .accountNo(account.getAccountId())
-                .startDate(account.getLastUpdated().format(formatter).substring(0, 8))
-                .endDate(code.substring(0, 8))
-                .transactionType("A")
-                .orderByType("ASC")
-                .build();
-
-            SSAFYResponse response = openFeignUtil.getListOfHistory(request);
-
-            List<Map<String, Object>> rawList = (List<Map<String, Object>>) response.getResultData()
-                .get("list");
-
-            List<HistoryResponse> res = rawList.stream()
-                .map(map -> objectMapper.convertValue(map, HistoryResponse.class))
-                .toList();
-
-            for (HistoryResponse historyResponse : res) {
-                String dateTimeString =
-                    historyResponse.getTransactionDate() + historyResponse.getTransactionTime();
-                LocalDateTime dateTime = LocalDateTime.parse(dateTimeString, formatter);
-
-                int type;
-                String transactionType = historyResponse.getTransactionTypeName();
-                type = switch (transactionType) {
-                    case "입금" -> 1;
-                    case "입금(수시입출금)" -> 2;
-                    case "출금" -> 3;
-                    case "출금(수시입출금)" -> 4;
-                    default -> 0;
-                };
-
-                History history = History.builder()
-                    .transactionId(historyResponse.getTransactionUniqueNo())
-                    .createdDate(dateTime.toLocalDate())
-                    .createdTime(dateTime.toLocalTime())
-                    .transactionAccount(historyResponse.getTransactionAccountNo())
-                    .transactionTarget(historyResponse.getTransactionSummary())
-                    .transactionAmount(Long.valueOf(historyResponse.getTransactionBalance()))
-                    .transactionType(type)
-                    .transactionTypeName(historyResponse.getTransactionTypeName())
-                    .category(Category.NOT_DEFINED)
-                    .account(account)
-                    .AmountAfterTransaction(
-                        Long.valueOf(historyResponse.getTransactionAfterBalance())
-                    )
-                    .memo(historyResponse.getTransactionMemo())
-                    .build();
-                historyRepository.save(history);
-            }
+            historyRepository.save(history);
 
             account.setLastUpdated(LocalDateTime.now());
         }
@@ -199,45 +230,35 @@ public class AccountService {
     }
 
     public void getVerificationCode(User user, String accountNo) {
+        if (user == null || user.getUserKey() == null || user.getUserKey().isEmpty()) {
+            throw new LoveLedgerException(ErrorCode.FORBIDDEN_ACCESS);
+        }
+
         String code = generateCode();
         String apiName = "openAccountAuth";
-        SSAFYRequestHeader header = SSAFYRequestHeader.builder()
-            .apiName(apiName)
-            .apiServiceCode(apiName)
-            .transmissionDate(code.substring(0, 8))
-            .transmissionTime(code.substring(8, 14))
-            .institutionCode("00100")
-            .fintechAppNo("001")
-            .institutionTransactionUniqueNo(code)
-            .apiKey(apiKey)
-            .userKey(user.getUserKey())
-            .build();
+        SSAFYRequestHeader header = createRequestHeader(user, apiName, code);
+
         AccountAuthenticationRequest request = AccountAuthenticationRequest.builder()
             .header(header)
-            .authText("SSAFY") // 차후 수정 필요
+            .authText("SSAFY") // TODO : 차후 수정 필요
             .accountNo(accountNo)
             .build();
 
-        SSAFYResponse response = openFeignUtil.sendAccountAuthentication(request);
-        String status = (String) response.getResultData().get("status");
+        try {
+            SSAFYResponse response = openFeignUtil.sendAccountAuthentication(request);
+        } catch (Exception e) {
+            throw new LoveLedgerException(
+                ErrorCode.OPENFEIGN_FAILED
+            );
+        }
     }
 
+    @Transactional
     public void doVerification(User user, String accountNo, String authCode) {
         String code = generateCode();
         String apiName = "checkAuthCode";
-        System.out.println(accountNo);
 
-        SSAFYRequestHeader header = SSAFYRequestHeader.builder()
-            .apiName(apiName)
-            .apiServiceCode(apiName)
-            .transmissionDate(code.substring(0, 8))
-            .transmissionTime(code.substring(8, 14))
-            .institutionCode("00100")
-            .fintechAppNo("001")
-            .institutionTransactionUniqueNo(code)
-            .apiKey(apiKey)
-            .userKey(user.getUserKey())
-            .build();
+        SSAFYRequestHeader header = createRequestHeader(user, apiName, code);
 
         AccountAuthenticationRequest request = AccountAuthenticationRequest.builder()
             .header(header)
@@ -248,17 +269,33 @@ public class AccountService {
 
         SSAFYResponse response = openFeignUtil.getAccountAuthentication(request);
         String status = (String) response.getResultData().get("status");
-
-        if (status.equals("SUCCESS")) {
+        if (status.equals("SUCCESS") && accountRepository.findById(accountNo).isEmpty()) {
             Account account = Account.builder()
                 .accountId(accountNo)
                 .bankCode("00100")
                 .user(user)
                 .certedAt(LocalDateTime.now())
                 .build();
-
             accountRepository.save(account);
+        } else {
+            throw new LoveLedgerException(
+                ErrorCode.ACCOUNT_CANT_CREATED
+            );
         }
+    }
+
+    private SSAFYRequestHeader createRequestHeader(User user, String apiName, String code) {
+        return SSAFYRequestHeader.builder()
+            .apiName(apiName)
+            .apiServiceCode(apiName)
+            .transmissionDate(code.substring(0, 8))
+            .transmissionTime(code.substring(8, 14))
+            .institutionCode("00100")
+            .fintechAppNo("001")
+            .institutionTransactionUniqueNo(code)
+            .apiKey(apiKey)
+            .userKey(user.getUserKey())
+            .build();
     }
 
     public void getMemberInfo(User user) {
